docs: explain `useNativeAnimations` for use cases
In cases where users use `LayoutAnimation` to smooth transitions upon swiping an item (e.g. perhaps swipe left deletes this Swipeable from a list), the non-open version of the Swipeable will be animated out (as opposed to the open version). This ends up looking bad/amatuer. Turns out this can be fixed by simply setting `useNativeAnimations` to false in those cases so that RN's `LayoutAnimation` shows the opened version of the Swipeable exiting.